+++ sql/pg_clickhouse--0.10--0.11.sql
@@ -1,0 +1,2 @@
+-- Replaced by clickhouse_query() and clickhouse_perform()
+DROP FUNCTION clickhouse_raw_query(text, text);